moved 'usdlib.py' to pipeline
