+++ client/ayon_core/hosts/houdini/plugins/publish/collect_instances_usd_layered.py
@@ -2,7 +2,7 @@ import hou
 import pyblish.api
 from ayon_core.hosts.houdini.api import lib
 import ayon_core.hosts.houdini.api.usd as hou_usdlib
-import ayon_core.lib.usdlib as usdlib
+from ayon_core.pipeline import usdlib
 
 
 class CollectInstancesUsdLayered(pyblish.api.ContextPlugin):
